go/netwatch: allow handlers to update pairs of views
Since sometimes (actually usually) a packet will contain information
about a src, and some dsts. While the src is arguably more trustworthy,
the dst still contains useful information.

However, after experimentation it was shown to cause volatile changes
in the host fields. So we keep the threading, and punt on the execution
until those bugs are fixed.