go/netwatch: allow filtering events based on shell exit code
This opens the possibility of filtering with more complex scripts using
the structured data from events. We call it a printableEvent to make
this clear.

There's still a few bugs with the ports being lost, but I'm leaving that
for a future commit, as it already existed before this commit.