watch: add ARP scan events for hosts
Closes https://github.com/henrywallace/netwatch/issues/4

We implement a windowing frequency data-structure for many entries, not
using expiring callbacks, but instead just using a slice. We use binary
search to flush old events so that it's relatively fast.

We also add a new Activity type that tracks IsActive, LastSeen, etc.
which is similar to the pattern already used for Hosts and Ports. In
fact we should just migrate them to use this type. But we punt on that
for now, so as not to make this commit to complicated.

There is also a bug in this implementation wrt to watching from PCAP
files, where the callbacks won't work properly. But this bug already
exists for many other aspects of the watch package. So we don't consider
this making things any worse. We will fix that in a future commit as
well.